Kyle's optimization advice
I did not account for the cases where I see a light right away but on the right side. With the current logic, it will require me to spin all the way around.
Second, I could also start off by seeing the left of the two lights first. This will require an extra calculation between light angles to determine.